webgpu::wgpu_thread

Struct WGPU

pub(crate) struct WGPU {
    receiver: IpcReceiver<WebGPURequest>,
    sender: IpcSender<WebGPURequest>,
    pub(crate) script_sender: IpcSender<WebGPUMsg>,
    pub(crate) global: Arc<Global>,
    devices: Arc<Mutex<FxHashMap<DeviceId, DeviceScope>>>,
    error_command_encoders: FxHashMap<CommandEncoderId, String>,
    pub(crate) paint_api: CrossProcessPaintApi,
    pub(crate) webrender_external_image_id_manager: WebRenderExternalImageIdManager,
    pub(crate) wgpu_image_map: WebGpuExternalImageMap,
    pub(crate) poller: Poller,
    compute_passes: FxHashMap<ComputePassId, Pass<ComputePass>>,
    render_passes: FxHashMap<RenderPassId, Pass<RenderPass>>,
}

Fields
receiver: IpcReceiver<WebGPURequest>sender: IpcSender<WebGPURequest>script_sender: IpcSender<WebGPUMsg>global: Arc<Global>devices: Arc<Mutex<FxHashMap<DeviceId, DeviceScope>>>error_command_encoders: FxHashMap<CommandEncoderId, String>

This stores first error on command encoder, because wgpu does not invalidate command encoder object (this is also reused for invalidation of command buffers)

paint_api: CrossProcessPaintApiwebrender_external_image_id_manager: WebRenderExternalImageIdManagerwgpu_image_map: WebGpuExternalImageMappoller: Poller

Provides access to poller thread

compute_passes: FxHashMap<ComputePassId, Pass<ComputePass>>

Store compute passes

render_passes: FxHashMap<RenderPassId, Pass<RenderPass>>

Store render passes

Implementations

impl WGPU

pub(crate) fn new( receiver: IpcReceiver<WebGPURequest>, sender: IpcSender<WebGPURequest>, script_sender: IpcSender<WebGPUMsg>, paint_api: CrossProcessPaintApi, webrender_external_image_id_manager: WebRenderExternalImageIdManager, wgpu_image_map: WebGpuExternalImageMap, ) -> Self

pub(crate) fn run(&mut self)

fn maybe_dispatch_wgpu_error<E: Error + 'static>( &mut self, device_id: DeviceId, error: Option<E>, )

fn maybe_dispatch_error(&mut self, device_id: DeviceId, error: Option<Error>)

Dispatches error (if there is any)

fn dispatch_error(&mut self, device_id: DeviceId, error: Error)

https://www.w3.org/TR/webgpu/#abstract-opdef-dispatch-error

fn encoder_record_error<U, T: Debug>( &mut self, encoder_id: CommandEncoderId, result: &Result<U, T>, )

impl WGPU

pub(crate) fn create_context( &self, context_id: WebGPUContextId, size: DeviceIntSize, buffer_ids: ArrayVec<BufferId, PRESENTATION_BUFFER_COUNT>, )

pub(crate) fn set_image_key( &self, context_id: WebGPUContextId, image_key: ImageKey, )

pub(crate) fn get_image( &self, context_id: WebGPUContextId, pending_texture: Option<PendingTexture>, sender: IpcSender<SharedSnapshot>, )

pub(crate) fn present( &self, context_id: WebGPUContextId, pending_texture: Option<PendingTexture>, size: Size2D<u32>, canvas_epoch: Epoch, )

Read the texture to the staging buffer, map it to CPU memory, and update the image in WebRender when complete.

fn texture_download( &self, texture_id: TextureId, encoder_id: CommandEncoderId, staging_buffer: StagingBuffer, config: ContextConfiguration, callback: impl FnOnce(StagingBuffer) + Send + 'static, )

Copies data from provided texture using encoder_id to the provided StagingBuffer.

callback is guaranteed to be called.

Returns a StagingBuffer with the StagingBufferState::Mapped state on success or StagingBufferState::Available on failure.

pub(crate) fn destroy_context(&mut self, context_id: WebGPUContextId)